Move testBacklogConsumerCacheReads to a separate test class
- it doesn't make sense to have this test in MessageDispatchThrottlingTest
